turkey-yacht-monitor/bot/requirements.txt
python-telegram-bot==21.6
requests>=2.31
